server/pyproject.toml
[tool.poetry]
name = "eventmanager-server"
version = "0.1.0"
description = "EventManager 서버 - 피부 진단 설문 및 AI 분류 서비스"
authors = ["EventManager Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

# 핵심 웹 프레임워크
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
starlette = "^0.27.0"
pydantic = "^2.12.3"

# AI 서비스
openai = "^1.57.4"

# 데이터베이스
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"

# 인증 및 보안
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^5.0.0"
cryptography = "^46.0.3"

# 유틸리티
python-dotenv = "^1.1.1"
python-multipart = "^0.0.6"
supabase = "^2.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.0.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
